added limit for auto filter values
added limit of 100 to auto filter values, because otherwise the tables can hang when enormous amounts of data have to be retrieved. this feature in general should be limited in use, when a lot of filter values exist. a better possibility would be to allow typeahead searching of values for those types of fields